Move CLI files
