+++ cli/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "osmforcities-cli",
+  "version": "0.1.0",
+  "description": "The command line interface for OSM for Cities",
+  "scripts": {
+    "cli": "TZ=UTC node src/index.js",
+    "lint": "eslint src"
+  },
+  "author": "Vitor George <vitor.george@gmail.com>",
+  "license": "MIT",
+  "type": "module",
+  "dependencies": {
+    "cli-progress": "^3.11.2",
+    "commander": "^9.4.0",
+    "knex": "^2.4.2",
+    "osmtogeojson": "^3.0.0-beta.4",
+    "pg": "^8.7.3",
+    "winston": "^3.9.0",
+    "winston-daily-rotate-file": "^4.7.1"
+  },
+  "devDependencies": {
+    "@fast-csv/parse": "^4.3.6",
+    "date-fns": "^2.30.0",
+    "dotenv": "^10.0.0",
+    "eslint": "^8.22.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.2.4",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.30.1",
+    "esm": "^3.2.25",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
+    "http-server": "^14.0.0",
+    "node-fetch": "^3.2.10",
+    "p-limit": "^4.0.0",
+    "simple-git": "^3.12.0"
+  }
+}
