+++ cli/src/helpers/curl-download.js
