+++ cli/src/helpers/pbf-is-empty.js
