+++ package.json
@@ -2,38 +2,12 @@
   "name": "osm-for-cities",
   "version": "0.1.0",
   "description": "A platform for distributing OpenStreetMap data presets for cities",
-  "scripts": {
-    "cli": "TZ=UTC node cli/index.js",
-    "lint": "eslint cli"
-  },
+  "scripts": {},
   "author": "Vitor George <vitor.george@gmail.com>",
   "license": "MIT",
   "type": "module",
-  "dependencies": {
-    "cli-progress": "^3.11.2",
-    "commander": "^9.4.0",
-    "knex": "^2.4.2",
-    "osmtogeojson": "^3.0.0-beta.4",
-    "pg": "^8.7.3",
-    "winston": "^3.9.0",
-    "winston-daily-rotate-file": "^4.7.1"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@fast-csv/parse": "^4.3.6",
-    "date-fns": "^2.29.3",
-    "dotenv": "^10.0.0",
-    "eslint": "^8.22.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.4",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.30.1",
-    "esm": "^3.2.25",
-    "execa": "^5.1.1",
-    "fs-extra": "^10.0.0",
-    "http-server": "^14.0.0",
-    "node-fetch": "^3.2.10",
-    "p-limit": "^4.0.0",
-    "simple-git": "^3.12.0"
+    "date-fns": "^2.30.0"
   }
 }
